ansible/playbooks/roles/llm_rag/files/requirements.txt
psycopg2-binary
flask
langchain
langchain-text-splitters
pdfminer.six
tiktoken
pgvector
requests
anthropic
openai
